LPS-84602 Move LayoutSetPrototypeTable to 7.1.x since it is intended to be used in this version and there is no need to be backported to 7.0.x
